Reorganized Frontend File Functions
