+++ client/src/pages/Home.tsx
@@ -13,7 +13,7 @@ import {
 
 import auth from "@/utils/auth";
 import { useState, useEffect } from "react";
-import { fetchFiles } from "@/services/fetchFiles";
+import { fetchFiles } from "@/services/file/fetchFiles";
 import { promptAI } from "@/services/promptAI";
 
 import GeoComp from "@/components/GeoSphere";
